Add parity check script for chapter detection and various improvements
- Introduced `parity_check_chapter_detect.py` to compare Rust and Python chapter detection implementations for consistency.
- Enhanced `_print_timing` function in `parity_check_chapter_topic_blocks.py` for improved readability.
- Removed unused `post_guid` from `AdClassifier` in `ad_classifier.py`.
- Added debug logging in `chapter_ad_detector.py` for fallback detection.
- Implemented clamping of chapter end times in `chapter_reader.py` to handle malformed ID3 frames.
- Updated `chapter_writer.py` to conditionally use Rust for writing chapters based on the audio flag.
- Added validation for chapter payloads in `rust_sidecar.py` to reject invalid data.
- Created unit tests for segment indexing and chapter validation in `test_chapter_fallback.py` and `test_chapter_rust_sidecar.py`.
- Improved test coverage for chapter writing behavior in `test_chapter_writer.py`.

+++ scripts/parity_check_chapter_topic_blocks.py
@@ -118,7 +118,7 @@ def _print_timing(label: str, timings_ms: list[float]) -> None:
         return
     sorted_t = sorted(timings_ms)
     median = statistics.median(sorted_t)
-    p95_idx = max(0, int(round(0.95 * (len(sorted_t) - 1))))
+    p95_idx = max(0, round(0.95 * (len(sorted_t) - 1)))
     p95 = sorted_t[p95_idx]
     mn = sorted_t[0]
     mx = sorted_t[-1]
